Reset volume on boot
Although the secondary microcontroller remembers the volume, it doesn't apply it on boot.
We therefore need to set it once we receive the response to the get_volume call we make on boot in order to avoid it defaulting to a setting which may not be what the user wants.
This also fixes the "first button press doesn't trigger chime" issue / closes #13.