Convert to an external component for easier usage
Converts to an external component, for easier usage with recent ESPHome versions.  The chime and button sensors have been removed, and events have been augmented with a "device" property which is populated with the ESPHome device name in order to allow disambiguation where multiple doorbells are connected to a single Home Assistant instance.